+++ Java/OOP/phoneSimulator/Main.java
@@ -1,0 +1,20 @@
+package com.company;
+
+public class Main {
+
+    public static void main(String[] args) {
+	People people1 = new People("Ahmet",1);
+	People people2 = new People("Temha",2);
+
+	people1.add(2);
+    people1.call(2);
+
+
+
+
+
+
+
+
+    }
+}
