+++ Java/OOP/phoneSimulator/foe.java
@@ -1,0 +1,12 @@
+package com.company;
+
+public class foe extends People {
+    foe(String Name, int ID) {
+        super(Name, ID);
+    }
+
+
+    public static void talk() {
+        System.out.println("Bugger OFF!");
+    }
+}
