--- phoneSimulator/bestFriend.java
@@ -1,13 +1,0 @@
-package com.company;
-
-public class bestFriend extends People{
-
-    bestFriend(String Name, int ID) {
-        super(Name, ID);
-    }
-
-    public static void talk() {
-        System.out.println("How are you my best friend ?");
-    }
-
-}
